moving from tmp to client/server/shared folders
