+++ client/__main__.py
@@ -1,22 +1,12 @@
 import asyncio
-import network
 import pygame
 
+from .client import nw
 
-HOST_ADDR = ('192.168.1.31', 12345)
-
-def on_receive(data):
-    print(f'Received {data}')
-
-def on_remote_close():
-    print('Connection closed by server')
-
-nw = network.create(HOST_ADDR, on_receive, on_remote_close)
 pygame.init()
 screen = pygame.display.set_mode((800, 600))
 pygame.display.set_caption('Client')
 clock = pygame.time.Clock()
-
 
 async def main():
     nw.start()
@@ -34,6 +24,5 @@ async def main():
         pygame.display.update()
         await asyncio.sleep(0)
         clock.tick(60)
-
 
 asyncio.run(main())
